pyproject.toml
[tool.poetry]
name = "note-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "note_service"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-crispy-forms = "^1.14.0"
django-filter = "^22.1"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
